ALL: fix compile warnings across platforms
- Replace %zu with %llu and (unsigned long long) cast in Log calls for MinGW compatibility
- Add [[fallthrough]] attribute in WindowedApp switch case for gcc -Wimplicit-fallthrough
- Fix signed/unsigned mismatch in VENGIFormatTest expected indices array

+++ src/modules/scenegraph/SceneGraph.cpp
@@ -1277,7 +1277,7 @@ SceneGraph::MergeResult SceneGraph::merge(bool skipHidden) const {
 	}
 	voxel::RawVolume *mergedVolume = new voxel::RawVolume(mergedRegion);
 	if (mergedVolume->voxels() == nullptr) {
-		Log::error("Failed to allocate merged volume (%zu bytes)", voxel::RawVolume::size(mergedRegion));
+		Log::error("Failed to allocate merged volume (%llu bytes)", (unsigned long long)voxel::RawVolume::size(mergedRegion));
 		delete mergedVolume;
 		return MergeResult{};
 	}
